# backend/tests/contract/openapi_v1_snapshot.json
{
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   },
   "OAuth2PasswordBearer": {
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/v1/auth/login"
     }
    },
    "type": "oauth2"
   }
  }
 },
 "info": {
  "contact": {
   "email": "support@valuadis.et",
   "name": "ValuAdis Support",
   "url": "https://valuadis.et/"
  },
  "description": "Ethiopian Property Valuation Platform API with PostGIS spatial support",
  "license": {
   "name": "MIT",
   "url": "https://opensource.org/licenses/MIT"
  },
  "title": "ValuAdis API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/vehicles/": {
   "get": {
    "description": "Get user's vehicles with optional filtering\n\nReturns a list of vehicles belonging to the current user with optional filters.",
    "operationId": "get_user_vehicles_api_v1_vehicles__get",
    "parameters": [
     {
      "description": "Number of records to skip",
      "in": "query",
      "name": "skip",
      "required": false,
      "schema": {
       "default": 0,
       "description": "Number of records to skip",
       "minimum": 0,
       "title": "Skip",
       "type": "integer"
      }
     },
     {
      "description": "Maximum number of records to return",
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 100,
       "description": "Maximum number of records to return",
       "maximum": 1000,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "description": "Filter by vehicle make",
      "in": "query",
      "name": "make",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Filter by vehicle make",
       "title": "Make"
      }
     },
     {
      "description": "Filter by vehicle model",
      "in": "query",
      "name": "model",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Filter by vehicle model",
       "title": "Model"
      }
     },
     {
      "description": "Filter by vehicle year",
      "in": "query",
      "name": "year",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "description": "Filter by vehicle year",
       "title": "Year"
      }
     },
     {
      "description": "Filter by region",
      "in": "query",
      "name": "region",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Filter by region",
       "title": "Region"
      }
     },
     {
      "description": "Filter by valuation status",
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "description": "Filter by valuation status",
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/VehicleResponse"
         },
         "title": "Response Get User Vehicles Api V1 Vehicles  Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get User Vehicles",
    "tags": [
     "vehicles"
    ]
   },
   "post": {
    "description": "Create a new vehicle record\n\nArgs:\n    vehicle_data: Vehicle creation data\n    \nReturns the created vehicle record.",
    "operationId": "create_vehicle_api_v1_vehicles__post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VehicleCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VehicleResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Vehicle",
    "tags": [
     "vehicles"
    ]
   }
  },
  "/api/v1/vehicles/statistics/summary": {
   "get": {
    "description": "Get vehicle statistics summary for the current user\n\nReturns statistics including total vehicles, total value, and breakdowns.",
    "operationId": "get_vehicle_statistics_api_v1_vehicles_statistics_summary_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Vehicle Statistics",
    "tags": [
     "vehicles"
    ]
   }
  },
  "/api/v1/vehicles/{vehicle_id}": {
   "delete": {
    "description": "Delete a vehicle record\n\nArgs:\n    vehicle_id: Vehicle ID\n    \nReturns success message if vehicle is deleted.",
    "operationId": "delete_vehicle_api_v1_vehicles__vehicle_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "vehicle_id",
      "required": true,
      "schema": {
       "title": "Vehicle Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Delete Vehicle",
    "tags": [
     "vehicles"
    ]
   },
   "get": {
    "description": "Get a specific vehicle by ID\n\nArgs:\n    vehicle_id: Vehicle ID\n    \nReturns the vehicle record if found and belongs to current user.",
    "operationId": "get_vehicle_api_v1_vehicles__vehicle_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "vehicle_id",
      "required": true,
      "schema": {
       "title": "Vehicle Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VehicleResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Vehicle",
    "tags": [
     "vehicles"
    ]
   },
   "put": {
    "description": "Update a vehicle record\n\nArgs:\n    vehicle_id: Vehicle ID\n    vehicle_data: Updated vehicle data\n    \nReturns the updated vehicle record.",
    "operationId": "update_vehicle_api_v1_vehicles__vehicle_id__put",
    "parameters": [
     {
      "in": "path",
      "name": "vehicle_id",
      "required": true,
      "schema": {
       "title": "Vehicle Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/VehicleUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VehicleResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Update Vehicle",
    "tags": [
     "vehicles"
    ]
   }
  },
  "/api/v1/vehicles/{vehicle_id}/latest-valuation": {
   "get": {
    "description": "Get the latest valuation for a specific vehicle\n\nArgs:\n    vehicle_id: Vehicle ID\n    \nReturns the most recent valuation for the specified vehicle.",
    "operationId": "get_latest_vehicle_valuation_api_v1_vehicles__vehicle_id__latest_valuation_get",
    "parameters": [
     {
      "in": "path",
      "name": "vehicle_id",
      "required": true,
      "schema": {
       "title": "Vehicle Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VehicleValuationResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Latest Vehicle Valuation",
    "tags": [
     "vehicles"
    ]
   }
  },
  "/api/v1/vehicles/{vehicle_id}/valuation": {
   "post": {
    "description": "Create a vehicle valuation\n\nArgs:\n    vehicle_id: Vehicle ID\n    valuation_data: Optional valuation data (will auto-calculate if not provided)\n    \nReturns the created vehicle valuation.",
    "operationId": "create_vehicle_valuation_api_v1_vehicles__vehicle_id__valuation_post",
    "parameters": [
     {
      "in": "path",
      "name": "vehicle_id",
      "required": true,
      "schema": {
       "title": "Vehicle Id",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "$ref": "#/components/schemas/VehicleValuationCreate"
         },
         {
          "type": "null"
         }
        ],
        "title": "Valuation Data"
       }
      }
     }
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/VehicleValuationResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Create Vehicle Valuation",
    "tags": [
     "vehicles"
    ]
   }
  },
  "/api/v1/vehicles/{vehicle_id}/valuations": {
   "get": {
    "description": "Get all valuations for a specific vehicle\n\nArgs:\n    vehicle_id: Vehicle ID\n    \nReturns a list of valuations for the specified vehicle.",
    "operationId": "get_vehicle_valuations_api_v1_vehicles__vehicle_id__valuations_get",
    "parameters": [
     {
      "in": "path",
      "name": "vehicle_id",
      "required": true,
      "schema": {
       "title": "Vehicle Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/VehicleValuationResponse"
         },
         "title": "Response Get Vehicle Valuations Api V1 Vehicles  Vehicle Id  Valuations Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "summary": "Get Vehicle Valuations",
    "tags": [
     "vehicles"
    ]
   }
  },
...
 },
...
}